fix: ensure that error is passed correctly to bugsnag
If the error was passed in any other argument than the first, the stack of the error object would be in metadata in bugsnag and in the same way as if the error object was sent as the first argument.
This leads to bugsnag showing the wrong stacktrace on the list-view of bugs.

To mitigate this we check if the info object has a stack-property. If so, we assume it's an error and new up an error with the same stacktrace.